requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
psycopg2-binary==2.9.9
Werkzeug==3.0.0
